% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{autoLISOMAP}
\alias{autoLISOMAP}
\title{autoLISOMAP: Build L-ISOMAP with automatically L1 and L2 size to build KNN graph}
\usage{
autoLISOMAP(Z, peaks = NULL, ndim = 3, L1 = NULL, L2 = NULL, N = 2)
}
\arguments{
\item{Z}{matrix, the PCA results}

\item{peaks}{peaks vector, the selected density peaks nodes}

\item{ndim}{integer, (default 2), set the dimensions of ISOMAP}

\item{L1}{integer, (default NULL),the kNN size for peaks}

\item{L2}{integer, (default NULL),the kNN size for all cells}

\item{N}{integer, (default 1), if N > 1, do parallel}
}
\value{
X, the ISOMAP matrix
}
\description{
autoLISOMAP: Build L-ISOMAP with automatically L1 and L2 size to build KNN graph
}
