% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{buildG_UMAP_PCA}
\alias{buildG_UMAP_PCA}
\title{getMSTLeafRootISO: create spanning tree with specified root and leaves}
\usage{
buildG_UMAP_PCA(
  uY,
  Z,
  dsM,
  ck = 8,
  ksize = 4,
  cluid = cluid,
  root = rt,
  leaves = list(a = 6, b = 8, c = 15, d = 14, e = 11, f = 13, g = 12)
)
}
\arguments{
\item{uY}{UMAP/t-SNE matrix}

\item{Z}{pca matrix}

\item{dsM}{distance matrix of clusters}

\item{ck}{integer, used to build knn graph for umap}

\item{ksize}{integer, default is 2, to build knn graph for clusters}

\item{cluid}{vector, the cluster id}

\item{root}{integer, default is 1, the root cluster}

\item{leaves}{list, specify the leave groups}
}
\value{
graph, g1, the trajectory.
}
\description{
getMSTLeafRootISO: create spanning tree with specified root and leaves
}
\examples{

}
