% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{calDensity}
\alias{calDensity}
\title{calDensity: Estimate the density of t-SNE}
\usage{
calDensity(tZ, fixed = TRUE, merge = FALSE, ksize = 50)
}
\arguments{
\item{tZ}{matrix, the t-SNE results}

\item{fixed}{logical, (default TRUE), Either 'fixed' or 'adaptive',
corresponding to a kernel estimator with fixed or adaptive bandwidths respectively. See kepdf().}

\item{ksize}{integer, (default 50), it is used to build KNN}
}
\value{
a list with three elements : pdf,peaks,valley
 pdf, the results of kepdf, users can plot the 3D density plot using pdf. See kepdf()
 peaks, the peaks nodes
 valley, all valley nodes
}
\description{
calDensity: Estimate the density of t-SNE
}
