% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{calSimilarity}
\alias{calSimilarity}
\title{calSimilarity: calculate the similarity among the clusters in PCA or ISOMAP space.}
\usage{
calSimilarity(X, Z, cluid, type = 1)
}
\arguments{
\item{X}{matrix, the ISOMAP matrix}

\item{Z}{matrix, the PCA matrix}

\item{cluid}{vector, the cluster id}

\item{type}{integer, (default 1), 1: X (ISOMAP) is used for silhouette index  ,
2: Z is used for silhouette index
3: Z is used for Pearson correlation instead of silhouette index}
}
\value{
list()
   list, rootDist: the similarity of clusters, cenDist: the distance of clusters
}
\description{
calSimilarity: calculate the similarity among the clusters in PCA or ISOMAP space.
}
