% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{cmcluster}
\alias{cmcluster}
\title{cmcluster: Build clustering based KNN graph and community clustering}
\usage{
cmcluster(
  Z,
  ksize = 2,
  cmmethod = "louvain",
  stepN = 4,
  clutrialN = 10,
  seed = 1
)
}
\arguments{
\item{Z}{matrix, the PCA results}

\item{ksize}{integer, (default 2), neighbor size to build knn graph}

\item{cmmethod}{string, (default louvain), user can also try fast_greedy method (cluster_fast_greedy),
walktrap ( cluster_walktrap() ) and infomap ( cluster_infomap())}

\item{stepN}{integer, used in walktrap method cluster_walktrap()}

\item{clutrialN}{integer, (default 10), for infomap method cluster_infomap()}
}
\value{
list
        knng, knn graph
        cluid, clustering results
}
\description{
cmcluster: Build clustering based KNN graph and community clustering
}
