% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{denknnCluT}
\alias{denknnCluT}
\title{denknnCluT: Do hierarchical clustering based on density of t-SNE}
\usage{
denknnCluT(denRes, tZ, cmethod1 = "ward", cmethod2 = "ward")
}
\arguments{
\item{tZ}{matrix, the t-SNE results}

\item{desRes}{list, the results of calDensity()}
}
\value{
memb, The clustering id
}
\description{
denknnCluT: Do hierarchical clustering based on density of t-SNE
}
