% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{distGraphClu}
\alias{distGraphClu}
\title{distGraphClu: Compute the distance among the clusters in knn graph}
\usage{
distGraphClu(Z, cluid, ksize = 4)
}
\arguments{
\item{Z}{matrix, the selected PCs}

\item{cluid}{vector, cell clustering results}

\item{ksize}{integer, neighbor siez to build knn graph}
}
\value{
matrix, the mean graph distance between each two clusters in graph
}
\description{
distGraphClu: Compute the distance among the clusters in knn graph
}
\examples{
    dsM = distGraphClu(Z, cluid,ksize=4)

}
