% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{doLmISOMAP}
\alias{doLmISOMAP}
\title{doLmISOMAP: Do landmark ISOMAP based on peaks and valleys}
\usage{
doLmISOMAP(Z, ndim = 2, peaks, valley, landMark = TRUE)
}
\arguments{
\item{Z}{matrix, the PCA results}

\item{ndim}{integer, (default 2), set the dimensions of ISOMAP}

\item{peaks}{vector, the selected density peaks nodes}

\item{valley}{vector, the selected density valley nodes}

\item{landMark}{logical, (default TRUE), whether use landmark, using landmark can save a lots of time}
}
\value{
X, the ISOMAP data
}
\description{
doLmISOMAP: Do landmark ISOMAP based on peaks and valleys
}
