% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{doLmISOMAPTemp}
\alias{doLmISOMAPTemp}
\title{doLmISOMAPTemp: Do landmark ISOMAP using the cluster centers as peaks}
\usage{
doLmISOMAPTemp(
  Z,
  ndim = 2,
  peaks,
  Lsize = 50,
  landMark = TRUE,
  clustering = FALSE
)
}
\arguments{
\item{Z}{matrix, the PCA results}

\item{ndim}{integer, (default 2), set the dimensions of ISOMAP}

\item{peaks}{vector, the selected density peaks nodes}

\item{landMark}{logical, (default TRUE), whether use landmark, using landmark can save a lots of time}

\item{valley}{vector, the selected density valley nodes}
}
\value{
X, the ISOMAP data
}
\description{
doLmISOMAPTemp: Do landmark ISOMAP using the cluster centers as peaks
}
