% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{drawMarkerValue}
\alias{drawMarkerValue}
\title{drawMarkerValue: plot marker genes expression among all clusters}
\usage{
drawMarkerValue(sd, memb, min.exp = 0.05, mean.expressing.only = F, title = "")
}
\arguments{
\item{sd}{a matrix of gene expression: row: marker genes, column: cells}

\item{memb}{The clusteirng id}

\item{min.exp}{Default 0.05. Minimum proportion of expressing cells (0-1) to be shown on the plot}

\item{mean.expressing.only}{(Logical) Should mean expression value exclude cells with no expression}

\item{title}{The title of the plot}
}
\value{

}
\description{
drawMarkerValue: plot marker genes expression among all clusters
}
\examples{

}
