% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ds_p2OB}
\alias{ds_p2OB}
\title{Calculate the distance form point A to edge OB}
\usage{
ds_p2OB(A, O, B)
}
\arguments{
\item{A}{point}

\item{O}{a point of edge OB}

\item{B}{a point of edge OB}
}
\value{
a vector of distance:
   length of edge AP
   length of edge OP
   length of edge OB
   the cosin value of the degree of edge OA and OB
}
\description{
Calculate the distance form point A to edge OB
}
\examples{
A=c(2,1) or A=c(1,1) or A=c(1,0)
B=c(2,0)
O=c(0,0)
ds_p2OB(A,O,B)
}
