% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{getCellSM}
\alias{getCellSM}
\title{getCellSM: the matrix of cell number from each cell type in each cluster}
\usage{
getCellSM(cluL, cellft)
}
\arguments{
\item{cluL}{cluster id}

\item{cellft}{cell annotations}
}
\value{
the matrix of cell number from each cell type in each cluster
}
\description{
getCellSM: the matrix of cell number from each cell type in each cluster
}
\examples{

}
