% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCenMST}
\alias{getCenMST}
\title{Get the minimum spanning tree}
\usage{
getCenMST(Y)
}
\arguments{
\item{Y}{The data matrix to build the MST}
}
\value{
the igraph object
}
\description{
Get the minimum spanning tree
}
