% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{getCluNeighbors}
\alias{getCluNeighbors}
\title{getCluNeighbors: Compute the neighbors number for each two clusters, if the number is larger than 0,
                 compute mean distance between two clusters.}
\usage{
getCluNeighbors(Z, cluid, ksize = 4)
}
\arguments{
\item{Z}{matrix, the selected PCs}

\item{cluid}{vector, cell clustering results}

\item{ksize}{integer, neighbor siez to build knn graph}
}
\value{
matrix, the mean graph distance between each two clusters in graph
}
\description{
getCluNeighbors: Compute the neighbors number for each two clusters, if the number is larger than 0,
                 compute mean distance between two clusters.
}
\examples{
    dsM = getCluNeighbors(Z, cluid,ksize=4)

}
