% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{getLayout}
\alias{getLayout}
\title{getLayout: get layout for the trajectory in 2D}
\usage{
getLayout(mst, root = 1)
}
\arguments{
\item{mst}{igraph, the trajectory}

\item{root}{Default 1}
}
\value{
igraph.layout
}
\description{
getLayout: get layout for the trajectory in 2D
}
\examples{

}
