% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{getLeafRootG}
\alias{getLeafRootG}
\title{getLeafRootG: create spanning tree with specified root and leaves}
\usage{
getLeafRootG(
  rootDist,
  cenDist,
  subcluid,
  L2 = 2,
  root = 1,
  localLeaf = list(a = c(9, 12, 13), b = 17, c = c(2, 6), d = 3, e = c(4, 5), f = 14)
)
}
\arguments{
\item{rootDist}{similariy matrix of clusters}

\item{cenDist}{distance matrix of clusters}

\item{subcluid}{vector, the cluster id}

\item{L2}{integer, default is 2, knn size}

\item{root}{integer, default is 1, the root cluster}

\item{localLeaf}{list, specify the leave groups}
}
\value{
list
     list(g1=g1,edList=edList)
}
\description{
getLeafRootG: create spanning tree with specified root and leaves
}
\examples{

}
