% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{getLeidenCluster}
\alias{getLeidenCluster}
\title{getLeidenCluster: apply leiden community clustering}
\usage{
getLeidenCluster(graph0, seed = 1)
}
\arguments{
\item{graph0}{graph object, input}
}
\value{
partition, vector, clustering results
}
\description{
getLeidenCluster: apply leiden community clustering
}
