% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{getMSTLeafRootISO}
\alias{getMSTLeafRootISO}
\title{getMSTLeafRootISO: create spanning tree with specified root and leaves}
\usage{
getMSTLeafRootISO(
  rootDist,
  cenDist,
  cluid,
  L2 = 2,
  root = 1,
  leaves = list(a = c(9, 12, 13), b = 17, c = c(2, 6), d = 3, e = c(4, 5), f = 14)
)
}
\arguments{
\item{rootDist}{similariy matrix of clusters}

\item{cenDist}{distance matrix of clusters}

\item{cluid}{vector, the cluster id}

\item{L2}{integer, default is 2, knn size}

\item{root}{integer, default is 1, the root cluster}

\item{leaves}{list, specify the leave groups}

\item{X}{ISOMAP matrix}
}
\value{
list
   list(g5=g5,rootDist=rootDist,w=sum(E(g5)$weight))
}
\description{
getMSTLeafRootISO: create spanning tree with specified root and leaves
}
\examples{

}
