% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{getMSTLeafRootISO_3}
\alias{getMSTLeafRootISO_3}
\title{getMSTLeafRootISO_3: create spanning tree with specified root and leaves}
\usage{
getMSTLeafRootISO_3(
  neigDist,
  cenDist,
  cluid,
  L2 = 2,
  root = 1,
  leaves = list(a = c(9, 12, 13, "linear"), b = 17, c = c(2, 6), d = 3, e = c(4, 5), f
    = 14)
)
}
\arguments{
\item{neigDist}{neighbor distance matrix of clusters}

\item{cenDist}{distance matrix of cluster centers}

\item{cluid}{vector, the cluster id}

\item{L2}{integer, default is 2, knn size}

\item{root}{integer, default is 1, the root cluster}

\item{leaves}{list, specify the leave groups, for each leave group, make it as a vector,
For example,a=c(9,12,13,"linear"), b=c(2,5, "mst"), c=c(7,8,9,"parallel")}
}
\value{
list
   list(g5=g5,neigDist=neigDist,w=sum(E(g5)$weight))
}
\description{
getMSTLeafRootISO_3: create spanning tree with specified root and leaves
}
\examples{

}
