% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{getMSTLeafRootISOMulroot}
\alias{getMSTLeafRootISOMulroot}
\title{getMSTLeafRootISOMulroot: create spanning tree with specified root and leaves}
\usage{
getMSTLeafRootISOMulroot(
  neigDist,
  cenDist,
  cluid,
  L2 = 2,
  rootL = 1,
  leaves = list(a = c(9, 12, 13, "linear"), b = 17, c = c(2, 6), d = 3, e = c(4, 5), f
    = 14)
)
}
\arguments{
\item{neigDist}{neighbor distance matrix of clusters}

\item{cenDist}{distance matrix of cluster centers}

\item{cluid}{vector, the cluster id}

\item{L2}{integer, default is 2, knn size}

\item{rootL}{integer vector, default is 1, the tree will connect the root cluster in linear mode
For example, rootL = c(1,2,3), the root edges will be 1->2->3}

\item{leaves}{list, specify the leave groups, for each leave group, make it as a vector,
For example,a=c(9,12,13,"linear"), b=c(2,5, "mst"), c=c(7,8,9,"parallel")}
}
\value{
list
   list(g5=g5,neigDist=neigDist,w=sum(E(g5)$weight))
}
\description{
getMSTLeafRootISOMulroot: create spanning tree with specified root and leaves
}
\examples{

}
