% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{getMSTLR}
\alias{getMSTLR}
\title{getMSTLR: create spanning tree with specified root and leaves}
\usage{
getMSTLR(
  g1,
  rootDist,
  cenDist,
  cluid,
  root = 1,
  leaves = list(a = c(9, 12, 13), b = 17, c = c(2, 6), d = 3, e = c(4, 5), f = 14)
)
}
\arguments{
\item{g1}{input graph}

\item{rootDist}{similariy matrix of clusters}

\item{cenDist}{distance matrix of clusters}

\item{cluid}{vector, the cluster id}

\item{root}{integer, default is 1, the root cluster}

\item{leaves}{list, specify the leave groups}
}
\value{
g1, graph
}
\description{
getMSTLR: create spanning tree with specified root and leaves
}
\examples{

}
