% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{kNNDist}
\alias{kNNDist}
\title{kNNDist: create KNN graph}
\usage{
kNNDist(rootDist, ks = 2)
}
\arguments{
\item{rootDist}{similariy matrix of clusters}

\item{ks}{integer, default is 2}
}
\value{
graph
   g1
}
\description{
kNNDist: create KNN graph
}
\examples{

}
