% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log2Filter}
\alias{log2Filter}
\title{Do log2 and filtering genes and cells}
\usage{
log2Filter(data, dolog = FALSE, mingenev = 1, cellPor = 0.01, cvth = 1)
}
\arguments{
\item{data}{matrix, the input data with row (genes) and column (cells)}

\item{dolog}{logical, whether to do log2, default FAlSE, not do log2}

\item{mingenev}{number, if the maximum expression value for the gene lower than the threshold,
the gene is filttered.}

\item{cellPor}{number, if the proportion of non-zero genes of the cell is lower than the threshold,
the cell is filttered}

\item{cvth}{number, if the variance of the gene is lower than the threshold,
the gene is filttered}
}
\value{
the log2 and filtered data
}
\description{
Do log2 and filtering genes and cells
}
