% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{mergeCluster}
\alias{mergeCluster}
\title{mergeCluster: merge clusters given specified cluster id}
\usage{
mergeCluster(mc = c(2, 4), cluid)
}
\arguments{
\item{mc}{vector the cluster id to merge}

\item{cluid}{vector, cell clustering results}
}
\value{
vector, the new merged cluster id
}
\description{
mergeCluster: merge clusters given specified cluster id
}
\examples{
    newCluid = getCluNeighbors(mc=c(2,4), cluid)

}
