% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{newOrder}
\alias{newOrder}
\title{newOrder: Infer pseduo time based on minimum spanning tree (MST) and given root in ISOMAP space}
\usage{
newOrder(X, Y, dp_mst = NULL, cluid, root = 1)
}
\arguments{
\item{X, }{the ISOMAP data;}

\item{Y, }{the cluster centers}

\item{dp_mst, }{the tree for mapping the cells. If it is null, use the clusters to build MST}

\item{cluid, }{the cluster id}

\item{root, }{the root node of MST}
}
\value{
ptime, matrix contians the pseudotime and distance from cell to the tree
}
\description{
newOrder: Infer pseduo time based on minimum spanning tree (MST) and given root in ISOMAP space
}
