% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{pcaL}
\alias{pcaL}
\title{pcaL: Do PCA and plot the PCs importance}
\usage{
pcaL(data, threshold = 0.001, nDim = NULL, center = TRUE, scale = FALSE)
}
\arguments{
\item{data}{a matrix with genes * cells}

\item{threshold}{numeric, (default 3), select the PCs based on proportion of variance of PCs}

\item{nDim}{numeric, (default 1:20), Set the top principle components(PCs) number. If it is not NULL, the threshold will not be used.}

\item{center}{logical, (default TRUE), it is used in PCA to do center for data}

\item{scale}{logical, (default FALSE), it is used in PCA to do scale for data}
}
\value{
list()
   a list with three elements : pca_out and Z
}
\description{
pcaL: Do PCA and plot the PCs importance
}
\examples{
 load(data);
           reducedData <- pcaL(data)


}
