% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{pcaL1}
\alias{pcaL1}
\title{pcaL1: Do PCA and select the number of PCs}
\usage{
pcaL1(
  data,
  pvalue = 0.05,
  pcaSel = 1,
  sel = 1,
  center = TRUE,
  scale = FALSE,
  ith = 0.01
)
}
\arguments{
\item{data}{a matrix with genes * cells}

\item{pvalue}{numeric, (default 0.05), select the PCs based on significant value for normal distribution}

\item{pcaSel}{numeric, (default 1 or 2), if pcaSel=1, matrix size < 10^8  109.78s;
if sel=2, For matrix size < 10^10}

\item{sel}{numeric, (default 1 or 2), if sel=1, use normal distibution to select top PCs;
if sel=2, do linear model to select top PCs from the top 100 PCs; it is faster than sel=1.
And the number of PCs selected are often larger than sel=1.}

\item{center}{logical, (default TRUE), it is used in PCA to do center for data}

\item{scale}{logical, (default FALSE), it is used in PCA to do scale for data}
}
\value{
list()
   a list with three elements : pca_out and the number of PCs selected
}
\description{
pcaL1: Do PCA and select the number of PCs
}
\examples{
 load(data);
           reducedData <- pcaL1(data) # For matrix size < 10^8
           reducedData <- pcaL1(data,sel=2) # For matrix size < 10^10


}
