% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{pieTreeRoot}
\alias{pieTreeRoot}
\title{pieTreeRoot: Visaulize the trajectory in pie tree}
\usage{
pieTreeRoot(
  mst,
  root = 1,
  ctyM,
  cellN,
  logSize = FALSE,
  legendFigWidth = c(2, 4),
  vs = 100,
  ew = 1,
  seed = 1,
  cex = 2,
  lbd = 1,
  lgs = 2,
  title = "Cells",
  colorname = "Spectral",
  showLab = FALSE
)
}
\arguments{
\item{mst}{igraph, the trajectory}

\item{root}{integer,  the root cluster id}

\item{ctyM}{cell type matrix}

\item{cellN}{vector, the names of clusters}

\item{logSize}{Default is FALSE, to do log on the cell number to reduce the cirdius of pie}

\item{legendFigWidth}{Default is c(2,4), 2 is the width of legend, 4 is the width of figure.
User can change them to adjust the layout of legend and figure}

\item{vs}{Default 100}

\item{ew}{Default 1 the label size}

\item{seed}{1}

\item{cex}{2}

\item{lbd}{1}

\item{lgs}{2}

\item{title}{string}

\item{colorname}{colorname: rownames(brewer.pal.info): "BrBG"     "PiYG"     "PRGn"     "PuOr"     "RdBu"     "RdGy"     "RdYlBu"   "RdYlGn"   "Spectral" "Accent"   "Dark2"
 "Paired"   "Pastel1"  "Pastel2"  "Set1"     "Set2"     "Set3"     "Blues"    "BuGn"     "BuPu"     "GnBu"     "Greens"
 "Greys"    "Oranges"  "OrRd"     "PuBu"     "PuBuGn"   "PuRd"     "Purples"  "RdPu"     "Reds"     "YlGn"     "YlGnBu"
"YlOrBr"   "YlOrRd"}

\item{showLab}{FALSE, show labels}
}
\value{

}
\description{
pieTreeRoot: Visaulize the trajectory in pie tree
}
\examples{

}
