% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{plot3DTree}
\alias{plot3DTree}
\title{plot3DTree: Visaulize trees on the 3D ISOMAP}
\usage{
plot3DTree(
  X,
  cluid,
  pointsize = 2,
  labelsize = 20,
  lineW = 5,
  colorname = "Spectral",
  titles = "3D ISOMAP"
)
}
\arguments{
\item{X}{3D ISOMAP:row samples, col:3D}

\item{cluid}{The clusteirng id}

\item{pointsize}{Default 2 for point size}

\item{labelsize}{Default 20. For enlarge the label size in the legend}

\item{lineW}{Default 5. line width in the MST}

\item{colorname}{colorname: rownames(brewer.pal.info): "BrBG"     "PiYG"     "PRGn"     "PuOr"     "RdBu"     "RdGy"     "RdYlBu"   "RdYlGn"   "Spectral" "Accent"   "Dark2"
 "Paired"   "Pastel1"  "Pastel2"  "Set1"     "Set2"     "Set3"     "Blues"    "BuGn"     "BuPu"     "GnBu"     "Greens"
 "Greys"    "Oranges"  "OrRd"     "PuBu"     "PuBuGn"   "PuRd"     "Purples"  "RdPu"     "Reds"     "YlGn"     "YlGnBu"
"YlOrBr"   "YlOrRd"}

\item{titles}{The title of the plot}
}
\value{

}
\description{
plot3DTree: Visaulize trees on the 3D ISOMAP
}
\examples{

}
