% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{plot3DTreeCells}
\alias{plot3DTreeCells}
\title{plot3DTreeCells: Visaulize the trajectory in 3D ISOMAP}
\usage{
plot3DTreeCells(
  X,
  peaks = NULL,
  cluid,
  celltype,
  mst = NULL,
  opcacity = 0.7,
  pointsize = 2,
  labelsize = 20,
  lineW = 5,
  colorname = "Spectral",
  titles = "ISOMAP"
)
}
\arguments{
\item{X}{matrix, the 3d ISOMAP}

\item{peaks}{Default is NULL, use the cluster centers as the vertexes.}

\item{cluid}{vector, the cluster id}

\item{celltype}{vector, the cell type annotation}

\item{mst}{igraph, the trajectory}

\item{opcacity}{Default 0.7}

\item{pointsize}{Default 2}

\item{labelsize}{Default 20}

\item{lineW}{5}

\item{colorname}{colorname: rownames(brewer.pal.info): "BrBG"     "PiYG"     "PRGn"     "PuOr"     "RdBu"     "RdGy"     "RdYlBu"   "RdYlGn"   "Spectral" "Accent"   "Dark2"
 "Paired"   "Pastel1"  "Pastel2"  "Set1"     "Set2"     "Set3"     "Blues"    "BuGn"     "BuPu"     "GnBu"     "Greens"
 "Greys"    "Oranges"  "OrRd"     "PuBu"     "PuBuGn"   "PuRd"     "Purples"  "RdPu"     "Reds"     "YlGn"     "YlGnBu"
"YlOrBr"   "YlOrRd"}

\item{titles}{FALSE, show labels}
}
\value{

}
\description{
plot3DTreeCells: Visaulize the trajectory in 3D ISOMAP
}
\examples{

}
