% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{plotAlls}
\alias{plotAlls}
\title{plotAlls: Visaulize 3D ISOMAP with different annotation}
\usage{
plotAlls(
  X = NULL,
  draw = 2,
  t = NULL,
  cluid = NULL,
  type = 1,
  size = 4,
  opacity = 1,
  labelsize = 20,
  titlename = "ISOMAP",
  colname = rownames(brewer.pal.info)[12],
  mcs = NULL
)
}
\arguments{
\item{X}{matrix}

\item{draw}{integer, 1:draw 2D in 3D space; 2: draw 2D 3: draw 3D}

\item{t}{Default for pseudotime annotation}

\item{cluid}{cluster id or other  annotation}

\item{type}{integer, 1: using t annotation 2: use cluid annotation 3. use cluid annotation and mcs colors}

\item{size}{Default 4. the cell size}

\item{opacity}{Default 1}

\item{labelsize}{Default 20. the label size}

\item{titlename}{string, the title}

\item{mcs}{The color vector specified by users}

\item{colorname}{colorname: rownames(brewer.pal.info): "BrBG"     "PiYG"     "PRGn"     "PuOr"     "RdBu"     "RdGy"     "RdYlBu"   "RdYlGn"   "Spectral" "Accent"   "Dark2"
 "Paired"   "Pastel1"  "Pastel2"  "Set1"     "Set2"     "Set3"     "Blues"    "BuGn"     "BuPu"     "GnBu"     "Greens"
 "Greys"    "Oranges"  "OrRd"     "PuBu"     "PuBuGn"   "PuRd"     "Purples"  "RdPu"     "Reds"     "YlGn"     "YlGnBu"
"YlOrBr"   "YlOrRd"}
}
\value{

}
\description{
plotAlls: Visaulize 3D ISOMAP with different annotation
}
\examples{

}
