% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{plotCellTree}
\alias{plotCellTree}
\title{plotCellTree: Visaulize the trajectory in 2D plot}
\usage{
plotCellTree(
  X,
  Y,
  cluid,
  mst = NULL,
  root = 1,
  colorby = "cluster",
  anno = NULL,
  colorName = rownames(brewer.pal.info)[22],
  height = 0.2,
  width = 0.2,
  pointSize = 0.2,
  edgeW = 0.5,
  textSize = 8,
  legendTextS = 20,
  legendColS = 10,
  hv = 0.5,
  vv = -1
)
}
\arguments{
\item{X}{matrix, the 3d ISOMAP}

\item{Y}{Default is the cluster centers as the vertexes.}

\item{cluid}{vector, the cluster id}

\item{mst}{igraph, the trajectory}

\item{root}{Default 1}

\item{colorby}{Default is "cluster"}

\item{anno}{other annotations}

\item{height}{Default 0.2}

\item{width}{0.2}

\item{pointSize}{Default 0.2}

\item{edgeW}{0.5}

\item{textSize}{Default 0.8}

\item{legendTextS}{20}

\item{legendColS}{10}

\item{hv}{0.5}

\item{vv}{-1}

\item{colorname}{colorname: rownames(brewer.pal.info): "BrBG"     "PiYG"     "PRGn"     "PuOr"     "RdBu"     "RdGy"     "RdYlBu"   "RdYlGn"   "Spectral" "Accent"   "Dark2"
 "Paired"   "Pastel1"  "Pastel2"  "Set1"     "Set2"     "Set3"     "Blues"    "BuGn"     "BuPu"     "GnBu"     "Greens"
 "Greys"    "Oranges"  "OrRd"     "PuBu"     "PuBuGn"   "PuRd"     "Purples"  "RdPu"     "Reds"     "YlGn"     "YlGnBu"
"YlOrBr"   "YlOrRd"}
}
\value{

}
\description{
plotCellTree: Visaulize the trajectory in 2D plot
}
\examples{

}
