% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{plotDens}
\alias{plotDens}
\title{plotDens: plot 3D density plot of tZ}
\usage{
plotDens(tZ, pdf, phi = 30, theta = 100, title = "data")
}
\arguments{
\item{tZ}{a matrix of tSNE}

\item{pdf}{An S4 object of kepdf-class with slots, see details in kepdf {pdfCluster}}

\item{phi}{phi, to adjust the degree of the up and down view of the figure in [0,360] range.}

\item{theta}{theta, to adjust the degree of the left and right view of the figure in [0,360] range.}
}
\value{
title
   The name of figure
}
\description{
plotDens: plot 3D density plot of tZ
}
\examples{

}
