% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{plotISOMST}
\alias{plotISOMST}
\title{plotISOMST: plot 2D cell trajectory on ISOMAP or PCA space}
\usage{
plotISOMST(
  X,
  type = c("ISOMAP", "PCA"),
  color_by = c("State", "Time", "Marker"),
  memb,
  discrete = TRUE,
  markername = "",
  x = 1,
  y = 2,
  Y,
  cenTree,
  show_tree = TRUE,
  show_cell_names = FALSE,
  cell_name_size = 3,
  pointsize = 2,
  lengendsize = 20,
  labelsize = 7
)
}
\arguments{
\item{X}{a matrix of ISOMAP or PCA}

\item{type}{"ISOMAP" or "PCA". It just changes the axis labels.}

\item{color_by}{"State", "Time", "Marker"}

\item{memb}{The clusteirng id}

\item{discrete}{To show whether the value of memb is discreate or continuous.}

\item{markername}{This the marker name when setting colory_by="Marker"}

\item{x}{The dimension of ISOMAP/PCA to visualize, default 1}

\item{y}{The dimension of ISOMAP/PCA to visualize, default 2}

\item{cenTree}{The minimum spanning tree}

\item{show_tree}{logical, (default TRUE), whether to draw the tree}

\item{show_cell_names}{logical, (default FALSE),Show cell name for each cell}

\item{cell_name_size}{Adjust the cell name size}

\item{pointsize}{Adjust the points size}

\item{lengendsize}{Adjust the lengend size}

\item{labelsize}{Adjust the label size for the verticle in the tree}
}
\value{

}
\description{
plotISOMST: plot 2D cell trajectory on ISOMAP or PCA space
}
\examples{

}
