% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{reduceByPT}
\alias{reduceByPT}
\title{reduceByPT: Do PCA and t-SNE for given data matrix}
\usage{
reduceByPT(data, threshold = 0.001, nDim = NULL, center = FALSE, scale = FALSE)
}
\arguments{
\item{data}{a matrix with genes * cells}

\item{threshold}{numeric, (default 3), select the PCs based on proportion of variance of PCs}

\item{nDim}{numeric, (default NULL), Set the top principle components(PCs) number. If it is not NULL, the threshold will not be used.}

\item{center}{logical, (default FALSE), it is used in PCA to do center for data}

\item{scale}{logical, (default FALSE), it is used in PCA to do scale for data}
}
\value{
list()
   a list with three elements : pca_out, tsne_out, and Z
}
\description{
reduceByPT: Do PCA and t-SNE for given data matrix
}
\examples{
 load(data);
           reducedData <- reduceByPT(data)


}
