% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{setLeignPath}
\alias{setLeignPath}
\title{setLeignPath: set the path of python package which installed the leidenalg package}
\usage{
setLeignPath(
  pyPath = "/Users/cheny48/miniconda3/bin/python",
  condaPath = "/Users/cheny48/miniconda3",
  libPath = "/Users/cheny48/miniconda3/lib/python3.8/site-packages"
)
}
\arguments{
\item{pyPath}{string, python command path}

\item{condaPath}{string, conda install path}

\item{libPath}{string, python library install path}
}
\description{
setLeignPath: set the path of python package which installed the leidenalg package
}
