% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{simpPTime}
\alias{simpPTime}
\title{simpPTime: Infer pseduo time based on minimum spanning tree (MST) and given root in ISOMAP space}
\usage{
simpPTime(X, cluid, cenTree = NULL, root = 1)
}
\arguments{
\item{X, }{the ISOMAP data;}

\item{cluid, }{the clustering id,}

\item{cenTree, }{the tree for mapping the cells. If it is null, use the clusters to build MST}

\item{root, }{the root node of MST}
}
\value{
Y, the nodes to build MST;
   cenTree, the MST;
   pt, pseduo time;
   cellDs, the list
}
\description{
simpPTime: Infer pseduo time based on minimum spanning tree (MST) and given root in ISOMAP space
}
