% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{tsneL}
\alias{tsneL}
\title{tsneL: Do tSNE}
\usage{
tsneL(Z, tdim = 2, center = FALSE, scale = FALSE, seeds = 1)
}
\arguments{
\item{Z}{PCA}

\item{tdim}{numeric, (default 2), how many dimensions for tSNE}

\item{center}{logical, (default FALSE), it is used in PCA to do center for data}

\item{scale}{logical, (default FALSE), it is used in PCA to do scale for data}
}
\value{
list()
   tsne_out
}
\description{
tsneL: Do tSNE
}
\examples{



}
