% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainMethods.R
\name{umpL}
\alias{umpL}
\title{umpL: Do UMAP}
\usage{
umpL(Z, seeds = 1)
}
\arguments{
\item{Z}{PCA}

\item{seeds}{numeric, (default 1), it is used to reproduce same UMAP}

\item{tdim}{numeric, (default 2), how many dimensions for tSNE}

\item{center}{logical, (default FALSE), it is used in PCA to do center for data}

\item{scale}{logical, (default FALSE), it is used in PCA to do scale for data}
}
\value{
list()
   tsne_out
}
\description{
umpL: Do UMAP
}
\examples{



}
